Renamed publisher package to distribution. Moved distribution and user package under targeting
