+++ android/src/main/java/com/adsbynimbus/openrtb/AndroidBidRequest.java
@@ -8,11 +8,11 @@ import androidx.collection.ArrayMap;
 
 import com.adsbynimbus.openrtb.impression.AndroidFormat;
 import com.adsbynimbus.openrtb.impression.Impression;
-import com.adsbynimbus.openrtb.publisher.App;
-import com.adsbynimbus.openrtb.user.Device;
-import com.adsbynimbus.openrtb.user.Regs;
-import com.adsbynimbus.openrtb.user.Source;
-import com.adsbynimbus.openrtb.user.User;
+import com.adsbynimbus.openrtb.targeting.distribution.App;
+import com.adsbynimbus.openrtb.targeting.user.Device;
+import com.adsbynimbus.openrtb.targeting.user.Regs;
+import com.adsbynimbus.openrtb.targeting.user.Source;
+import com.adsbynimbus.openrtb.targeting.user.User;
 
 import java.lang.annotation.Retention;
